Make setup instructions more clear
I misunderstood the setup instructions due to lack of experience with cypress plugins. The current README buries the instructions for adding the plugin to `cypress/plugins` way down in the "Details" section. For new Cypress devs like me, this gives the impression that it's not a required step for the plugin to work.

This change makes it clear that this step is required, and places example code directly in the README.

I created an issue based on my misunderstanding. Clarifying the README fixes #49.